+++ examples/consume/consumer.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"log"
 
-	"github.com/pkg/errors"
 	"github.com/streadway/amqp"
 )
 
@@ -27,7 +26,9 @@ func (c *Consumer) Declare(ctx context.Context, ch *amqp.Channel) error {
 		nil,            // arguments
 	)
 	if err != nil {
-		return errors.WithMessage(err, "failed to declare "+c.ExchangeName)
+		log.Printf("failed to declare exchange %v: %v", c.ExchangeName, err)
+
+		return err
 	}
 
 	_, err = ch.QueueDeclare(
@@ -36,10 +37,14 @@ func (c *Consumer) Declare(ctx context.Context, ch *amqp.Channel) error {
 		false,       // delete when unused
 		false,       // exclusive
 		false,       // no-wait
-		nil,         // arguments
+		amqp.Table{
+			"x-max-length": int64(40),
+		},
 	)
 	if err != nil {
-		return errors.WithMessage(err, "failed to declare "+c.QueueName)
+		log.Printf("failed to declare queue %v: %v", c.QueueName, err)
+
+		return err
 	}
 
 	err = ch.QueueBind(
@@ -50,7 +55,9 @@ func (c *Consumer) Declare(ctx context.Context, ch *amqp.Channel) error {
 		nil,            // arguments
 	)
 	if err != nil {
-		return errors.WithMessage(err, "failed to bind "+c.QueueName+" to "+c.ExchangeName)
+		log.Printf("failed to bind queue %v: %v", c.QueueName, err)
+
+		return err
 	}
 
 	return nil
@@ -66,7 +73,9 @@ func (c *Consumer) Consume(ctx context.Context, ch *amqp.Channel) error {
 		false, // global
 	)
 	if err != nil {
-		return errors.WithMessage(err, "failed to set qos")
+		log.Printf("failed to set qos: %v", err)
+
+		return err
 	}
 
 	msgs, err := ch.Consume(
@@ -79,7 +88,9 @@ func (c *Consumer) Consume(ctx context.Context, ch *amqp.Channel) error {
 		nil,          // args
 	)
 	if err != nil {
-		return errors.WithMessage(err, "failed to consume "+c.QueueName)
+		log.Printf("failed to consume %v: %v", c.QueueName, err)
+
+		return err
 	}
 
 	for {